[flang2] Cast int to size_t before comparison with sizeof; NFCI
The result of sizeof is frequently used in comparisons with signed int
variables. Casting the other operand in such comparisons to size_t
silences a number of -Wsign-compare warnings.

+++ tools/flang2/flang2exe/dinit.cpp
@@ -4278,7 +4278,7 @@ eval_transfer(CONST *arg, DTYPE dtype)
   /* Be sure we have enough space. */
   int need = (rsize > ssize ? rsize : ssize) * 2;
 
-  if (sizeof(value) < need) {
+  if (sizeof(value) < (size_t)need) {
     NEW(buffer, char, need);
     if (buffer == NULL)
       return CONST_ERR(dtype);
